Generalize fake with --mode argument
